move make_error_body
